STYLE: Move SlicerPathology scripted module into its own folder.
Consistent organization with one module per directory.

Also renamed the python module Util to SlicerPathologyUtil
