+++ SlicerPathology/Resources/Colors/SlicerPathology.csv
