+++ SlicerPathology/Resources/Icons/SlicerPathology-simple-small.png
