+++ SlicerPathology/Resources/Icons/SlicerPathology.png
