+++ SlicerPathology/Resources/Icons/SlicerPathology128px.png
