+++ SlicerPathology/Resources/Icons/SlicerPathologyExtensionLogo.png
