+++ SlicerPathology/Resources/Icons/SlicerPathologyLogo128px.png
