+++ SlicerPathology/Resources/Icons/icon-imageselection.png
