+++ SlicerPathology/Resources/Icons/icon-segmentation.png
