+++ SlicerPathology/Resources/Icons/icon-setup.png
