+++ SlicerPathology/Resources/Icons/icon-submission.png
