+++ SlicerPathology/SlicerPathologyUtil/mixins.py
